% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottom_legend.R
\name{bottom_legend}
\alias{bottom_legend}
\title{bottom_legend}
\usage{
bottom_legend()
}
\value{
A ggplot object with a graphical guide for the scale on the bottom.
}
\description{
bottom_legend
}
\examples{
data <- data.frame(x=factor(c('A','A','A','B','B','C','C')),y=c(10,20,30,40,50,60,70))
ggplot2::ggplot(data) + ggplot2::geom_violin(ggplot2::aes(x=x,y=y,fill=x)) + bottom_legend()
}
