% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\title{freq}
\usage{
freq(data, x)
}
\arguments{
\item{data}{A data.frame containing the data.}

\item{x}{A categorical variable (factor or character) to compute frequencies for.}
}
\value{
A data.frame with the following columns:
\itemize{
\item \code{modalites}: unique levels of the variable \code{x}.
\item \code{freq}: absolute frequencies of each level.
\item \code{freq_relative}: relative frequencies in percentages.
\item \code{cumfreq}: cumulative absolute frequencies.
\item \code{cumfreq_relative}: cumulative relative frequencies in percentages.
}
}
\description{
freq
}
\examples{
data <- data.frame(
id = factor(c(1,1,2,2,2,2,3,3,3,4,4,4,5,5,5),levels= c(1,2,3,4,5)))
freq(data,id)
}
