% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfishcol.R
\name{ggfishcol}
\alias{ggfishcol}
\title{ggfishcol}
\usage{
ggfishcol()
}
\value{
A ggplot object with a fill color depending on the species.
}
\description{
ggfishcol
}
\examples{
data <- data.frame(x=factor(c('SAT','BRO','PER')),y=c('12','15','28'))
ggplot2::ggplot(data) + ggplot2::geom_bar(ggplot2::aes(x=x,y=y,fill=x),stat="identity") + ggfishcol()
}
