% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfishline.R
\name{ggfishline}
\alias{ggfishline}
\title{ggfishline}
\usage{
ggfishline()
}
\value{
A ggplot object with a line color depending on the species.
}
\description{
ggfishline
}
\examples{
data <- data.frame(x=factor(c('SAT','BRO','PER')),y=c('12','15','28'))
ggplot2::ggplot(data) + ggplot2::geom_bar(ggplot2::aes(x=x,y=y,color=x),stat="identity") + ggfishline()
}
