% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggletter.R
\name{ggletter}
\alias{ggletter}
\title{ggletter}
\usage{
ggletter(data, x, y, test)
}
\arguments{
\item{data}{A data.frame containing the data.}

\item{x}{A categorical variable (factor).}

\item{y}{A numeric variable.}

\item{test}{The name of the statistical test.}
}
\value{
A ggplot object with the significative difference letter marker.
}
\description{
ggletter
}
\examples{
data <- data.frame(
x = factor(c("A","A","A","A","A","B","B","B","C","C","C","C","C")),
y = c(10,10,11,20,30,40,50,10,1600,1600,2100,4000,2500))
ggplot2::ggplot(data,ggplot2::aes(x=x,y=y)) + ggplot2::geom_violin() + ggletter(data,x,y,"kruskal")
}
