% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmean.R
\name{ggmean}
\alias{ggmean}
\title{ggmean}
\usage{
ggmean(size = 4.5, color = "black", fill = "white", shape = 21)
}
\arguments{
\item{size}{Size of the point representing the average.}

\item{color}{Color of the point outline.}

\item{fill}{Dot fill color.}

\item{shape}{Point shape.}
}
\value{
A ggplot stat_summary addition to display the average.
}
\description{
ggmean
}
\examples{
data <- data.frame(x=factor(c('A','A','A','B','B','C','C')),y=c(10,20,30,40,50,60,70))
ggplot2::ggplot(data,ggplot2::aes(x=x,y=y)) + ggplot2::geom_violin() + ggmean()
}
