% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnex.R
\name{ggnex}
\alias{ggnex}
\title{ggnex}
\usage{
ggnex(data, x)
}
\arguments{
\item{data}{A data.frame containing the data.}

\item{x}{A categorical variable in the data.frame.}
}
\value{
A ggplot object with modified x-axis labels.
}
\description{
ggnex
}
\examples{
data <- data.frame(x=factor(c('A','A','A','B','B','C')))
ggplot2::ggplot(data) + ggplot2::geom_bar(ggplot2::aes(x=x)) + ggnex(data,x)
}
