% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnin.R
\name{ggnin}
\alias{ggnin}
\title{ggnin}
\usage{
ggnin(data, x, y)
}
\arguments{
\item{data}{A data.frame containing the data.}

\item{x}{A categorical variable (factor) used for grouping.}

\item{y}{A numeric variable used for the y-axis.}
}
\value{
A ggplot object with the count of observations displayed.
}
\description{
ggnin
}
\examples{
data <- data.frame(x=factor(c('A','A','A','B','B','C','C')),y=c(10,20,30,40,50,60,70))
ggplot2::ggplot(data) + ggplot2::geom_violin(ggplot2::aes(x=x, y=y)) + ggnin(data,x,y)
}
