% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtitlex.R
\name{ggtitlex}
\alias{ggtitlex}
\title{ggtitlex}
\usage{
ggtitlex()
}
\value{
A ggplot object with a graphic charter.
}
\description{
ggtitlex
}
\examples{
data <- data.frame(x=factor(c('A','A','A','B','B','C','C')),y=c(10,20,30,40,50,60,70))
ggplot2::ggplot(data) + ggplot2::geom_violin(ggplot2::aes(x=x, y=y)) + ggtitlex()
}
