% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_legend.R
\name{right_legend}
\alias{right_legend}
\title{right_legend}
\usage{
right_legend()
}
\value{
A ggplot object with a graphical guide for the scale on the right.
}
\description{
right_legend
}
\examples{
data <- data.frame(x=factor(c('A','A','A','B','B','C','C')),y=c(10,20,30,40,50,60,70))
ggplot2::ggplot(data) + ggplot2::geom_violin(ggplot2::aes(x=x, y=y, fill=x)) + right_legend()
}
