% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat}
\alias{stat}
\title{stat}
\usage{
stat(...)
}
\arguments{
\item{...}{Additional arguments (currently unused).}
}
\value{
A statistical summary of the data.
}
\description{
stat
}
\examples{
data <- data.frame(y=c(10,20,30,40,50,60,70,80,90,100,110,120,130,140,150,160))
stat(data$y)
}
