% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat2.R
\name{stat2}
\alias{stat2}
\title{stat2}
\usage{
stat2(data, x, y)
}
\arguments{
\item{data}{A data.frame containing the data.}

\item{x}{A numeric variable.}

\item{y}{A categorical variable (factor).}
}
\value{
A statistical summary of the data.
}
\description{
stat2
}
\examples{
data <- data.frame(x=factor(c('A','A','A','A','A','B','B','B','B','B','B')),y=c(10,20,30,40,50,60,70,80,90,100,110))
stat2(data,x=y,y=x)
}
