% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat3.R
\name{stat3}
\alias{stat3}
\title{stat3}
\usage{
stat3(data, x, y)
}
\arguments{
\item{data}{A data.frame containing the data.}

\item{x}{A numeric variable.}

\item{y}{A categorical variable (factor).}
}
\value{
A statistical summary of the data.
}
\description{
stat3
}
\examples{
data <- data.frame(x=factor(c('A','A','A','A','A','B','B','B','B','B','B','C','C','C','C','C')),y=c(10,20,30,40,50,60,70,80,90,100,110,120,130,140,150,160))
stat3(data,x=y,y=x)
}
